more renames, adjust tests
