+++ conbench/api/__init__.py
@@ -13,7 +13,7 @@ rule: Callable = api.add_url_rule
 
 from ._errors import *  # noqa
 from .auth import *  # noqa
-from .benchmarks import *  # noqa
+from .results import *  # noqa
 from .commits import *  # noqa
 from .compare import *  # noqa
 from .contexts import *  # noqa
