+++ conbench/app/index.py
@@ -7,7 +7,7 @@ from sqlalchemy import select
 
 from ..app import rule
 from ..app._endpoint import AppEndpoint, authorize_or_terminate
-from ..app.benchmarks import RunMixin
+from ..app.results import RunMixin
 from ..config import Config
 from ..db import Session
 from ..entities.run import Run
